+++ backend/controllers/recipeController.js
@@ -9,16 +9,16 @@ const {
 
 exports.getAllRecipesHandler = async (req, res, next) => {
   try {
-    const searchString = Object.entries(req.query)
-      .map(([key, value]) => `${key} ILIKE '%${value}%'`)
-      .join(" AND ");
+    // const searchString = Object.entries(req.query)
+    //   .map(([key, value]) => `${key} ILIKE '%${value}%'`)
+    //   .join(" AND ");
 
     const filter = {
-      limit: req.query.limit || 12,
-      offset: (req.query.page - 1) * req.query.limit || 0,
+      limit: +req.query.limit || 12,
+      offset: (+req.query.page - 1) * +req.query.limit || 0,
       sortBy: req.query.sortBy || "title",
       order: req.query.order || "asc",
-      searchString,
+      // searchString,
     };
 
     const recipes = await getAllRecipes(filter);
